@@ -2,16 +2,20 @@
 
 buildscript {
     ext {
-        kotlin_version = '1.3.72'
+        kotlin_version = '1.5.0'
+
+        //kotlin_version = '1.4.10'
     }
     repositories {
-        jcenter()
+        mavenCentral()
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.0.1'
-        classpath 'com.google.gms:google-services:4.3.3'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath 'com.android.tools.build:gradle:4.2.2'
+        classpath 'com.google.gms:google-services:4.3.8'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.21"
+
+        classpath 'com.google.dagger:hilt-android-gradle-plugin:2.37' //TODO Hilt step 1
         //classpath "com.android.databinding:dataBinder:1.0-rc1"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -20,7 +24,7 @@ buildscript {
 
 allprojects {
     repositories {
-        jcenter()
+        mavenCentral()
         google()
          maven {
              url 'https://maven.google.com'
