@@ -1,14 +1,11 @@
 package com.myapp.lexicon.service;
 
 
-import android.content.ContentValues;
-import android.content.Intent;
-import android.os.AsyncTask;
+import android.app.AlertDialog;
+import android.app.Dialog;
 import android.os.Bundle;
-import android.speech.tts.TextToSpeech;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.widget.Button;
@@ -16,446 +13,224 @@ import android.widget.CheckBox;
 import android.widget.CompoundButton;
 import android.widget.ImageButton;
 import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.AdView;
 import com.myapp.lexicon.R;
-import com.myapp.lexicon.database.DataBaseEntry;
-import com.myapp.lexicon.database.DatabaseHelper;
-import com.myapp.lexicon.database.GetEntriesFromDbAsync;
-import com.myapp.lexicon.database.GetStudiedWordsCount;
-import com.myapp.lexicon.database.UpdateDBEntryAsync;
-import com.myapp.lexicon.dialogs.WordsEndedDialog;
+import com.myapp.lexicon.ads.AdsViewModel;
+import com.myapp.lexicon.billing.BillingViewModel;
+import com.myapp.lexicon.database.Word;
 import com.myapp.lexicon.helpers.RandomNumberGenerator;
-import com.myapp.lexicon.main.MainActivityOnStart;
-import com.myapp.lexicon.main.SplashScreenActivity;
-import com.myapp.lexicon.settings.AppData;
-import com.myapp.lexicon.settings.AppSettings;
-
-import org.greenrobot.eventbus.EventBus;
+import com.myapp.lexicon.helpers.StringOperations;
+import com.myapp.lexicon.interfaces.IModalFragment;
+import com.myapp.lexicon.main.MainViewModel;
+import com.myapp.lexicon.main.SpeechViewModel;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Date;
+import java.util.List;
+import java.util.Locale;
 
 import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentActivity;
-import androidx.fragment.app.FragmentManager;
-
-import static com.myapp.lexicon.main.MainActivity.serviceIntent;
-import static com.myapp.lexicon.service.ServiceDialog.map;
-import static com.myapp.lexicon.service.ServiceDialog.speech;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.DialogFragment;
+import androidx.lifecycle.ViewModelProvider;
+import dagger.hilt.android.AndroidEntryPoint;
+import io.reactivex.disposables.CompositeDisposable;
 
 
-public class TestModalFragment extends Fragment
+@AndroidEntryPoint
+public class TestModalFragment extends DialogFragment
 {
-    private AppSettings appSettings;
-    private AppData appData;
+    public static final String TAG = TestModalFragment.class.getCanonicalName() + ".TAG";
+    public static IModalFragment iCallback;
+
     private TextView enTextView;
     private Button ruBtn1, ruBtn2;
-    private ImageView orderPlayIcon;
-    private TextView nameDictTV;
-    private TextView wordsNumberTV;
-    private ArrayList<DataBaseEntry> compareList;
+    private List<Word> compareList;
     private boolean wordIsStudied = false;
-    private boolean isWordsEnded = false;
-    private WordsEndedDialog endedDialog = null;
+    private static List<Integer> _counters = new ArrayList<>();
+    private Word[] words = new Word[0];
+
+    private MainViewModel viewModel;
+    private SpeechViewModel speechVM;
+    private AdsViewModel adsVM;
+    private final CompositeDisposable composite = new CompositeDisposable();
 
     public TestModalFragment()
     {
         // Required empty public constructor
     }
 
-    static TestModalFragment newInstance()
+    static TestModalFragment newInstance(@Nullable String json, List<Integer> counters, IModalFragment callback)
     {
-        return new TestModalFragment();
+        TestModalFragment fragment = new TestModalFragment();
+        _counters = counters;
+        iCallback = callback;
+        if (json != null && counters.size() > 0)
+        {
+            Bundle bundle = new Bundle();
+            bundle.putString(ServiceActivity.ARG_JSON, json);
+            fragment.setArguments(bundle);
+        }
+        return fragment;
     }
 
     @Override
     public void onCreate(Bundle savedInstanceState)
     {
         super.onCreate(savedInstanceState);
-        setRetainInstance(true);
-        if (getActivity() != null)
-        {
-            appSettings = new AppSettings(getActivity());
-            appData = AppData.getInstance();
-            appData.initAllSettings(getActivity());
-        } else
-        {
-            onDestroy();
-            onDetach();
-        }
+        viewModel = new ViewModelProvider(this).get(MainViewModel.class);
+        speechVM = new  ViewModelProvider(this).get(SpeechViewModel.class);
     }
 
+    @NonNull
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
+    public Dialog onCreateDialog(@Nullable Bundle savedInstanceState)
     {
-        View fragmentView = inflater.inflate(R.layout.s_test_modal_fragment, container, false);
-        enTextView = fragmentView.findViewById(R.id.en_text_view);
-        enTextView.setText("");
-        ruBtn1 = fragmentView.findViewById(R.id.ru_btn_1);
+        View dialogView = LayoutInflater.from(requireContext()).inflate(R.layout.s_test_modal_fragment, new LinearLayout(requireContext()), false);
+
+        AlertDialog dialog = new AlertDialog.Builder(requireContext()).setView(dialogView).create();
+
+        BillingViewModel billingVM = new ViewModelProvider(this).get(BillingViewModel.class);
+        adsVM = new ViewModelProvider(this).get(AdsViewModel.class);
+        billingVM.getNoAdsToken().observe(this, token -> {
+            if (token != null && token.isEmpty())
+            {
+                LinearLayout adLayout = dialogView.findViewById(R.id.adLayout);
+                if (adLayout != null)
+                {
+                    AdView mainBanner = adsVM.getMainBanner();
+                    adLayout.addView(mainBanner);
+                    mainBanner.loadAd(new AdRequest.Builder().build());
+                }
+            }
+        });
+
+        enTextView = dialogView.findViewById(R.id.en_text_view);
+        ruBtn1 = dialogView.findViewById(R.id.ru_btn_1);
         ruBtn1.setText("");
         ruBtn1_OnClick(ruBtn1);
-        ruBtn2 = fragmentView.findViewById(R.id.ru_btn_2);
+        ruBtn2 = dialogView.findViewById(R.id.ru_btn_2);
         ruBtn2.setText("");
         ruBtn2_OnClick(ruBtn2);
 
-
-        nameDictTV = fragmentView.findViewById(R.id.name_dict_tv_test_modal);
-        wordsNumberTV = fragmentView.findViewById(R.id.words_number_tv_test_modal);
-
-        ImageButton speakButton = fragmentView.findViewById(R.id.btn_sound_modal);
-        speakButton_OnClick(speakButton);
-
-        String currentDict = null;
-        try
+        Bundle arguments = getArguments();
+        if (arguments != null)
         {
-            currentDict = appSettings.getPlayList().get(appData.getNdict());
-        }
-        catch (IndexOutOfBoundsException e)
-        {
-            appData.setNdict(0);
-            appData.setNword(1);
-            if (appSettings.getPlayList().size() > 0)
+            String json = arguments.getString(ServiceActivity.ARG_JSON);
+            try
             {
-                currentDict = appSettings.getPlayList().get(appData.getNdict());
-            }
-        }
-
-        try
-        {
-            nameDictTV.setText(currentDict);
-            int orderPlay = appSettings.getOrderPlay();
-            switch (orderPlay)
-            {
-                case 0:
-                    getWordsFromDBbyOrder(currentDict);
-                    break;
-                case 1:
-                    getRandomWordsFromDB();
-                    break;
-            }
-
-        } catch (Exception e)
-        {
-            e.printStackTrace();
-        }
-
-        ImageButton btnClose = fragmentView.findViewById(R.id.modal_btn_close);
-        btnClose.setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
-            {
-                if (getActivity() != null)
+                if (json != null)
                 {
-                    getActivity().finish();
-                }
-            }
-        });
-
-        Button btnOpenApp = fragmentView.findViewById(R.id.btn_open_app);
-        btnOpenApp_OnClick(btnOpenApp);
-
-        Button btnStopService = fragmentView.findViewById(R.id.btn_stop_service);
-        btnStopService_OnClick(btnStopService);
-
-        checkStudied_OnCheckedChange((CheckBox) fragmentView.findViewById(R.id.check_box_studied));
-
-        orderPlayIcon = fragmentView.findViewById(R.id.order_play_icon_iv_test_modal);
-
-        return fragmentView;
-    }
-
-    @Override
-    public void onResume()
-    {
-        super.onResume();
-        if (appSettings.getOrderPlay() == 0)
-        {
-            orderPlayIcon.setImageResource(R.drawable.ic_repeat_white);
-        }
-        if (appSettings.getOrderPlay() == 1)
-        {
-            orderPlayIcon.setImageResource(R.drawable.ic_shuffle_white);
-        }
-    }
-
-    @Override
-    public void onStart()
-    {
-        super.onStart();
-    }
-
-    @Override
-    public void onStop()
-    {
-        super.onStop();
-    }
-
-    private void getWordsFromDBbyOrder(final String currentDict)
-    {
-        GetStudiedWordsCount getStudiedWordsCount = new GetStudiedWordsCount(getActivity(), currentDict, new GetStudiedWordsCount.GetCountListener()
-        {
-            int firstId = appData.getNword();
-
-            @Override
-            public void onTaskComplete(Integer[] resArray)
-            {
-                if (resArray != null && resArray.length > 1)
-                {
-                    final int maxCount = resArray[3];
-                    final int studiedCount = resArray[2];
-                    if (studiedCount == maxCount && getActivity() != null)
+                    words = StringOperations.getInstance().jsonToWord(json);
+                    if (words.length > 0)
                     {
-                        appSettings.removeItemFromPlayList(currentDict);
-                        int ndict = appData.getNdict() + 1;
-                        if (ndict > appSettings.getPlayList().size() - 1)
-                        {
-                            appSettings.setDictNumber(0);
-                            appSettings.setWordNumber(1);
-                        }
-                        getActivity().finish();
-                    }
-                    try
-                    {
-                        if (maxCount - studiedCount > 0)
-                        {
-                            GetEntriesFromDbAsync getEntriesFromDbAsync = new GetEntriesFromDbAsync(getActivity(), GetEntriesFromDbAsync.KEY_GET_TWO_DISTINCT_WORDS, currentDict, firstId, new GetEntriesFromDbAsync.GetEntriesListener()
-                            {
-                                @Override
-                                public void getEntriesListener(ArrayList<DataBaseEntry> entries)
-                                {
-                                    int wordsNumber = 0;
-                                    compareList = entries;
-                                    if (entries.size() == 1)
-                                    {
-                                        wordsNumber = entries.get(0).getRowId();
-                                        enTextView.setText(entries.get(0).getEnglish());
-                                        ruBtn1.setText(entries.get(0).getTranslate());
-                                        ruBtn2.setText(entries.get(0).getTranslate());
-                                        if (appData.getNword() + 1 <= studiedCount)
-                                        {
-                                            appData.setNword(appData.getNword() + 1);
-                                        }
-                                        else
-                                        {
-                                            appData.setNword(1);
-                                            if (appData.getNdict() + 1 > appSettings.getPlayList().size() - 1)
-                                            {
-                                                appData.setNdict(0);
-                                            }
-                                            else if (appData.getNdict() >= 0 && appData.getNdict() <  appSettings.getPlayList().size() - 1)
-                                            {
-                                                appData.setNdict(appData.getNdict() + 1);
-                                            }
-                                            else
-                                            {
-                                                appData.setNdict(appData.getNdict());
-                                            }
-                                        }
-                                    }
-                                    if (entries.size() == 2)
-                                    {
-                                        RandomNumberGenerator numberGenerator = new RandomNumberGenerator(2, (int) new Date().getTime());
-                                        int i = numberGenerator.generate();
-                                        int j = numberGenerator.generate();
-                                        wordsNumber = entries.get(0).getRowId();
-                                        enTextView.setText(entries.get(0).getEnglish());
-                                        ruBtn1.setText(entries.get(i).getTranslate());
-                                        ruBtn2.setText(entries.get(j).getTranslate());
+                        enTextView.setText(words[0].getEnglish());
+                        TextView nameDictTV = dialogView.findViewById(R.id.name_dict_tv_test_modal);
+                        nameDictTV.setText(words[0].getDictName());
 
-                                        if (appData.getNword() + 1 <= studiedCount)
-                                        {
-                                            appData.setNword(appData.getNword() + 1);
-                                        }
-                                        else
-                                        {
-                                            appData.setNword(1);
-                                            if (appData.getNdict() + 1 > appSettings.getPlayList().size() - 1)
-                                            {
-                                                appData.setNdict(0);
-                                            }
-                                            else if (appData.getNdict() >= 0 && appData.getNdict() <  appSettings.getPlayList().size() - 1)
-                                            {
-                                                appData.setNdict(appData.getNdict() + 1);
-                                            }
-                                            else
-                                            {
-                                                appData.setNdict(appData.getNdict());
-                                            }
-                                        }
-                                    }
-                                    if (entries.size() == 3)
-                                    {
-                                        RandomNumberGenerator numberGenerator = new RandomNumberGenerator(2, (int) new Date().getTime());
-                                        int i = numberGenerator.generate();
-                                        int j = numberGenerator.generate();
-
-                                        wordsNumber = entries.get(0).getRowId();
-                                        enTextView.setText(entries.get(0).getEnglish());
-                                        ArrayList<Button> buttons = new ArrayList<Button>()
-                                        {{
-                                            add(ruBtn1);
-                                            add(ruBtn2);
-                                        }};
-                                        buttons.get(i).setText(entries.get(0).getTranslate());
-                                        buttons.get(j).setText(entries.get(2).getTranslate());
-                                        appData.setNword(entries.get(1).getRowId());
-                                    }
-                                    if (entries.size() > 0)
-                                    {
-                                        try
-                                        {
-                                            String concatText = (wordsNumber + "").concat(" / ").concat(Integer.toString(maxCount)).concat(" " + getString(R.string.text_studied) + " " + studiedCount);
-                                            wordsNumberTV.setText(concatText);
-                                        } catch (Exception e)
-                                        {
-                                            e.printStackTrace();
-                                        }
-                                    }
-                                }
-                            });
-                            if (getEntriesFromDbAsync.getStatus() != AsyncTask.Status.RUNNING)
-                            {
-                                getEntriesFromDbAsync.execute();
-                            }
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        wordsNumberTV.setText("???");
-                    }
-                }
-            }
-        });
-        if (getStudiedWordsCount.getStatus() != AsyncTask.Status.RUNNING)
-        {
-            getStudiedWordsCount.execute();
-        }
-    }
-
-    private void getRandomWordsFromDB()
-    {
-        if (appSettings.getPlayList().size() > 0)
-        {
-            RandomNumberGenerator numberGenerator = new RandomNumberGenerator(appSettings.getPlayList().size(), (int) new Date().getTime());
-            int nDict = numberGenerator.generate();
-            final String tableName = appSettings.getPlayList().get(nDict);
-            if (tableName == null) return;
-
-            GetStudiedWordsCount getStudiedWordsCount = new GetStudiedWordsCount(getActivity(), tableName, new GetStudiedWordsCount.GetCountListener()
-            {
-                @Override
-                public void onTaskComplete(Integer[] resArray)
-                {
-                    if (resArray != null && resArray.length > 1)
-                    {
-                        final int totalWords = resArray[3];
-                        final int studiedWords = resArray[2];
-                        if (studiedWords == totalWords && getActivity() != null)
-                        {
-                            appSettings.removeItemFromPlayList(tableName);
-                            getActivity().finish();
-                        }
-
-                        GetEntriesFromDbAsync getEntriesFromDbAsync = new GetEntriesFromDbAsync(getActivity(), tableName, new GetEntriesFromDbAsync.GetEntriesListener()
-                        {
-                            @Override
-                            public void getEntriesListener(ArrayList<DataBaseEntry> entries)
-                            {
-                                compareList = entries;
-                                int wordsNumber = 0;
-                                if (entries.size() == 1)
-                                {
-                                    wordsNumber = entries.get(0).getRowId();
-                                    enTextView.setText(entries.get(0).getEnglish());
-                                    ruBtn1.setText(entries.get(0).getTranslate());
-                                    ruBtn2.setText(entries.get(0).getTranslate());
-                                }
-                                if (entries.size() > 1)
-                                {
-                                    RandomNumberGenerator numberGenerator = new RandomNumberGenerator(2, (int) new Date().getTime());
-                                    int i = numberGenerator.generate();
-                                    int j = numberGenerator.generate();
-                                    wordsNumber = entries.get(0).getRowId();
-                                    enTextView.setText(entries.get(0).getEnglish());
-                                    ruBtn1.setText(entries.get(i).getTranslate());
-                                    ruBtn2.setText(entries.get(j).getTranslate());
-                                }
-                                nameDictTV.setText(tableName);
-                                try
-                                {
-                                    String concatText = (wordsNumber + "").concat(" / ").concat(Integer.toString(totalWords)).concat(" " + getString(R.string.text_studied) + " " + studiedWords);
-                                    wordsNumberTV.setText(concatText);
-                                } catch (Exception e)
-                                {
-                                    e.printStackTrace();
-                                }
-                            }
+                        viewModel.getRandomWord(words[0]).observe(this, word -> {
+                            ArrayList<Word> listWords = new ArrayList<>();
+                            listWords.add(words[0]);
+                            listWords.add(word);
+                            RandomNumberGenerator numberGenerator = new RandomNumberGenerator(2, (int) new Date().getTime());
+                            int i = numberGenerator.generate();
+                            int j = numberGenerator.generate();
+                            ruBtn1.setText(listWords.get(i).getTranslate());
+                            ruBtn2.setText(listWords.get(j).getTranslate());
+                            compareList = listWords;
                         });
-                        if (getEntriesFromDbAsync.getStatus() != AsyncTask.Status.RUNNING)
-                        {
-                            getEntriesFromDbAsync.execute();
-                        }
                     }
                 }
-            });
-            if (getStudiedWordsCount.getStatus() != AsyncTask.Status.RUNNING)
-            {
-                getStudiedWordsCount.execute();
+                TextView wordsNumberTV = dialogView.findViewById(R.id.words_number_tv_test_modal);
+                if (_counters.size() >= 3)
+                {
+                    String concatText = (_counters.get(0) + "")
+                            .concat(" / ")
+                            .concat(_counters.get(1) + "")
+                            .concat("  " + getString(R.string.text_studied) + " " + _counters.get(2));
+                    wordsNumberTV.setText(concatText);
+                }
             }
+            catch (Exception e)
+            {
+                e.printStackTrace();
+            }
+
         }
+
+        ImageButton speakButton = dialogView.findViewById(R.id.btn_sound_modal);
+        speakButton.setOnClickListener(view -> {
+            String enText = enTextView.getText().toString();
+            if (!enText.equals(""))
+            {
+                speechVM.doSpeech(enText, Locale.US);
+            }
+        });
+
+        ImageButton btnClose = dialogView.findViewById(R.id.modal_btn_close);
+        btnClose.setOnClickListener(view -> requireActivity().finish());
+
+        Button btnOpenApp = dialogView.findViewById(R.id.btn_open_app);
+        btnOpenApp.setOnClickListener(view1 -> iCallback.openApp());
+
+        Button btnStopService = dialogView.findViewById(R.id.btn_stop_service);
+        btnStopService.setOnClickListener( view -> ((ServiceActivity)requireActivity()).stopAppService());
+
+        checkStudied_OnCheckedChange(dialogView.findViewById(R.id.check_box_studied));
+
+        ImageView orderPlayIcon = dialogView.findViewById(R.id.order_play_icon_iv_test_modal);
+
+        viewModel.getCountRepeat().observe(this, id -> {
+            if (id > 0) Toast.makeText(getActivity(), R.string.text_word_is_not_show, Toast.LENGTH_LONG).show();
+        });
+
+        if (dialog.getWindow() != null) dialog.getWindow().setBackgroundDrawableResource(R.drawable.rounded_corners_background);
+        return dialog;
     }
 
     private void ruBtn1_OnClick(View view)
     {
-        view.setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
+        view.setOnClickListener( v -> {
+            Button button = (Button) v;
+            String translate = button.getText().toString().toLowerCase();
+            String english = enTextView.getText().toString().toLowerCase();
+            boolean result = compareWords(compareList, english, translate);
+            if (result)
             {
-                Button button = (Button) view;
-                String trnslate = button.getText().toString().toLowerCase();
-                String english = enTextView.getText().toString().toLowerCase();
-                boolean result = compareWords(compareList, english, trnslate);
-                if (result)
-                {
-                    rightAnswerAnim(button);
-                }
-                else
-                {
-                    noRightAnswerAnim(button);
-                }
+                rightAnswerAnim(button);
+            }
+            else
+            {
+                noRightAnswerAnim(button);
             }
         });
     }
 
     private void ruBtn2_OnClick(View view)
     {
-        view.setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
+        view.setOnClickListener( v -> {
+            Button button = (Button) v;
+            String translate = button.getText().toString().toLowerCase();
+            String english = enTextView.getText().toString().toLowerCase();
+            boolean result = compareWords(compareList, english, translate);
+            if (result)
             {
-                Button button = (Button) view;
-                String trnslate = button.getText().toString().toLowerCase();
-                String english = enTextView.getText().toString().toLowerCase();
-                boolean result = compareWords(compareList, english, trnslate);
-                if (result)
-                {
-                    rightAnswerAnim(button);
-                }
-                else
-                {
-                    noRightAnswerAnim(button);
-                }
+                rightAnswerAnim(button);
+            }
+            else
+            {
+                noRightAnswerAnim(button);
             }
         });
     }
 
-    private boolean compareWords(ArrayList<DataBaseEntry> compareList, String english, String translate)
+    private boolean compareWords(List<Word> compareList, String english, String translate)
     {
         boolean result = false;
         if (compareList != null && compareList.size() > 0)
@@ -486,97 +261,16 @@ public class TestModalFragment extends Fragment
             @Override
             public void onAnimationStart(Animation animation)
             {
-                if (getActivity() != null)
+                viewModel.goForward(Arrays.asList(words));
+                button.setBackgroundResource(R.drawable.btn_for_test_modal_green);
+                button.setTextColor(requireContext().getResources().getColor(R.color.colorWhite));
+
+                if (wordIsStudied)
                 {
-                    button.setBackgroundResource(R.drawable.btn_for_test_modal_green);
-                    button.setTextColor(getActivity().getResources().getColor(R.color.colorWhite));
-                    if (wordIsStudied)
+                    if (words.length > 0)
                     {
-                        String dict = nameDictTV.getText().toString();
-                        DataBaseEntry entry = new DataBaseEntry(enTextView.getText().toString(), button.getText().toString(), "0");
-                        ContentValues values = new ContentValues();
-                        values.put(DatabaseHelper.COLUMN_Count_REPEAT, 0);
-                        UpdateDBEntryAsync updateDBEntryAsync = new UpdateDBEntryAsync(getActivity(), dict, values, "English = ? AND Translate = ?", new String[]{entry.getEnglish(), entry.getTranslate()}, new UpdateDBEntryAsync.IUpdateDBListener()
-                        {
-                            @Override
-                            public void updateDBEntry_OnComplete(int rows)
-                            {
-                                if (rows > 0 && getActivity() != null)
-                                {
-                                    Toast.makeText(getActivity(), R.string.text_word_is_not_show, Toast.LENGTH_LONG).show();
-                                    final String currentDict = nameDictTV.getText().toString();
-                                    GetStudiedWordsCount getCountWordsAsync = new GetStudiedWordsCount(getActivity(), currentDict, new GetStudiedWordsCount.GetCountListener()
-                                    {
-                                        @Override
-                                        public void onTaskComplete(Integer[] resArray)
-                                        {
-                                            if (resArray.length > 1)
-                                            {
-                                                int notStudied = resArray[0];
-                                                if (notStudied == 0)
-                                                {
-                                                    isWordsEnded = true;
-                                                    FragmentManager fragmentManager = getActivity().getSupportFragmentManager();
-                                                    Fragment fragmentByTag = fragmentManager.findFragmentByTag(WordsEndedDialog.TAG);
-                                                    if (fragmentByTag != null)
-                                                    {
-                                                        fragmentManager.beginTransaction().remove(fragmentByTag).commit();
-                                                    }
-
-                                                    endedDialog = WordsEndedDialog.getInstance(currentDict, new WordsEndedDialog.IWordEndedDialogResult()
-                                                    {
-                                                        @Override
-                                                        public void wordEndedDialogResult(int res)
-                                                        {
-                                                            switch (res)
-                                                            {
-                                                                case 0:
-                                                                    appSettings.removeItemFromPlayList(currentDict);
-                                                                    if (appSettings.getPlayList() == null || appSettings.getPlayList().size() == 0)
-                                                                    {
-                                                                        getActivity().stopService(serviceIntent);
-                                                                    }
-                                                                    getActivity().finish();
-                                                                    break;
-                                                                case 1:
-                                                                    ContentValues values = new ContentValues();
-                                                                    values.put(DatabaseHelper.COLUMN_Count_REPEAT, 1);
-                                                                    UpdateDBEntryAsync updateDBEntryAsync = new UpdateDBEntryAsync(getActivity(), currentDict, values, null, null, new UpdateDBEntryAsync.IUpdateDBListener()
-                                                                    {
-                                                                        @Override
-                                                                        public void updateDBEntry_OnComplete(int rows)
-                                                                        {
-                                                                            if (getActivity() != null)
-                                                                            {
-                                                                                getActivity().finish();
-                                                                            }
-                                                                        }
-                                                                    });
-                                                                    if (updateDBEntryAsync.getStatus() != AsyncTask.Status.RUNNING)
-                                                                    {
-                                                                        updateDBEntryAsync.execute();
-                                                                    }
-                                                                    break;
-                                                            }
-
-                                                        }
-                                                    });
-                                                    fragmentManager.beginTransaction().add(endedDialog, WordsEndedDialog.TAG).commit();
-                                                }
-                                            }
-                                        }
-                                    });
-                                    if (getCountWordsAsync.getStatus() != AsyncTask.Status.RUNNING)
-                                    {
-                                        getCountWordsAsync.execute();
-                                    }
-                                }
-                            }
-                        });
-                        if (updateDBEntryAsync.getStatus() != AsyncTask.Status.RUNNING)
-                        {
-                            updateDBEntryAsync.execute();
-                        }
+                        int wordId = words[0].get_id();
+                        viewModel.setCountRepeat(0, wordId, wordId);
                     }
                 }
             }
@@ -584,20 +278,9 @@ public class TestModalFragment extends Fragment
             @Override
             public void onAnimationEnd(Animation animation)
             {
-                if (getActivity() != null)
-                {
-                    button.setBackgroundResource(R.drawable.btn_for_test_modal_transp);
-                    button.setTextColor(getActivity().getResources().getColor(R.color.colorLightGreen));
-                    if (getActivity() != null && !isWordsEnded)
-                    {
-                        appData.saveAllSettings(getActivity());
-                        getActivity().finish();
-                    } else
-                    {
-                        onDestroy();
-                        onDetach();
-                    }
-                }
+                button.setBackgroundResource(R.drawable.btn_for_test_modal_transp);
+                button.setTextColor(requireContext().getResources().getColor(R.color.colorLightGreen));
+                requireActivity().finish();
             }
 
             @Override
@@ -644,60 +327,7 @@ public class TestModalFragment extends Fragment
         button.startAnimation(animNotRight);
     }
 
-    private void speakButton_OnClick(View view)
-    {
-        view.setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
-            {
-                if (speech == null || speech.isSpeaking())
-                {
-                    return;
-                }
-                String enText = enTextView.getText().toString();
-                if (!enText.equals(""))
-                {
-                    speech.speak(enText, TextToSpeech.QUEUE_ADD, map);
-                }
-            }
-        });
-    }
-
-    private void btnOpenApp_OnClick(View view)
-    {
-        view.setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
-            {
-                if (getActivity() != null)
-                {
-                    getActivity().startActivity(new Intent(getContext(), SplashScreenActivity.class));
-                    getActivity().finish();
-                    EventBus.getDefault().postSticky(new MainActivityOnStart(serviceIntent));
-                }
-            }
-        });
-    }
-
-    private void btnStopService_OnClick(Button button)
-    {
-        button.setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
-            {
-                FragmentActivity activity = getActivity();
-                if (activity != null)
-                {
-                    LexiconService.stopedByUser = true;
-                    EventBus.getDefault().post(new StopedServiceByUserEvent());
-                }
-            }
-        });
-    }
-
+    @SuppressWarnings("Convert2Lambda")
     private void checkStudied_OnCheckedChange(CheckBox checkBox)
     {
         checkBox.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener()
@@ -710,5 +340,11 @@ public class TestModalFragment extends Fragment
         });
     }
 
-
+    @Override
+    public void onDestroy()
+    {
+        super.onDestroy();
+        composite.dispose();
+        composite.clear();
+    }
 }
