@@ -12,14 +12,16 @@ import com.myapp.lexicon.helpers.RandomNumberGenerator;
 
 import java.util.Date;
 
+import androidx.core.content.ContextCompat;
+
 /**
  * TODO: AdapterViewFlipper: 5. адаптер для AdapterViewFlipper
  */
 
 public class FlipperAdapter extends BaseAdapter
 {
-    private Context context;
-    private int[] images;
+    private final Context context;
+    private final int[] images;
     private RandomNumberGenerator generator; // генератор случайных чисел из диапазона без повторяющихся значений
 
     FlipperAdapter(Context context, int[] images)
@@ -40,7 +42,7 @@ public class FlipperAdapter extends BaseAdapter
     @Override
     public Object getItem(int i)
     {
-        return context.getResources().getDrawable(images[i]);
+        return ContextCompat.getDrawable(context, images[i]);
     }
 
     @Override
