@@ -1,33 +1,18 @@
 package com.myapp.lexicon.main;
 
-import android.app.ActivityManager;
-import android.app.AlertDialog;
-import android.app.Service;
-import android.content.BroadcastReceiver;
+import android.app.NotificationManager;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.content.IntentFilter;
 import android.content.SharedPreferences;
 import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.provider.Settings;
-import android.speech.tts.TextToSpeech;
-import android.speech.tts.UtteranceProgressListener;
-import android.text.Editable;
-import android.text.InputType;
-import android.text.TextWatcher;
-import android.util.DisplayMetrics;
-import android.view.Gravity;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.WindowManager;
+import android.widget.Button;
 import android.widget.CheckBox;
-import android.widget.CompoundButton;
-import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
@@ -35,6 +20,8 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.AdView;
 import com.google.android.gms.ads.MobileAds;
 import com.google.android.material.navigation.NavigationView;
 import com.google.firebase.appindexing.Action;
@@ -44,84 +31,274 @@ import com.google.firebase.appindexing.builders.Actions;
 import com.myapp.lexicon.R;
 import com.myapp.lexicon.aboutapp.AboutAppFragment;
 import com.myapp.lexicon.addword.TranslateFragment;
+import com.myapp.lexicon.ads.AdsViewModel;
+import com.myapp.lexicon.billing.BillingViewModel;
 import com.myapp.lexicon.cloudstorage.StorageFragment2;
-import com.myapp.lexicon.database.DataBaseEntry;
-import com.myapp.lexicon.database.DataBaseQueries;
+import com.myapp.lexicon.database.AppDB;
+import com.myapp.lexicon.database.AppDao;
+import com.myapp.lexicon.database.AppDataBase;
 import com.myapp.lexicon.database.DatabaseHelper;
-import com.myapp.lexicon.database.GetTableListFragm;
+import com.myapp.lexicon.database.Word;
+import com.myapp.lexicon.databinding.ANavigMainBinding;
+import com.myapp.lexicon.dialogs.DictListDialog;
+import com.myapp.lexicon.dialogs.OrderPlayDialog;
+import com.myapp.lexicon.dialogs.RemoveDictDialog;
+import com.myapp.lexicon.helpers.AppBus;
 import com.myapp.lexicon.helpers.Share;
-import com.myapp.lexicon.playlist.PlayList;
+import com.myapp.lexicon.schedule.AlarmScheduler;
 import com.myapp.lexicon.service.LexiconService;
-import com.myapp.lexicon.settings.AppData;
-import com.myapp.lexicon.settings.AppSettings;
 import com.myapp.lexicon.settings.SettingsFragment;
 import com.myapp.lexicon.wordeditor.WordEditorActivity;
-import com.myapp.lexicon.wordstests.TestsActivity;
-
-import org.greenrobot.eventbus.EventBus;
+import com.myapp.lexicon.wordstests.OneOfFiveFragm;
+import com.myapp.lexicon.wordstests.TestFragment;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 
+import javax.inject.Inject;
+
+import androidx.activity.OnBackPressedCallback;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.AppCompatImageButton;
 import androidx.appcompat.widget.Toolbar;
 import androidx.core.view.GravityCompat;
 import androidx.drawerlayout.widget.DrawerLayout;
-import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.viewpager2.widget.ViewPager2;
+import dagger.hilt.android.AndroidEntryPoint;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.disposables.CompositeDisposable;
+import io.reactivex.disposables.Disposable;
+import io.reactivex.schedulers.Schedulers;
 
 
-public class MainActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener,
-        AppData.IDictNumChangeListener,
-        GetTableListFragm.OnTableListListener
+@AndroidEntryPoint
+public class MainActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener
 {
-    public DatabaseHelper databaseHelper;
-    private Intent addWordIntent;
-    private Intent wordEditorIntent;
-    private Intent testsIntent;
-    private Intent playListIntent;
-    private TextView textViewEn;
-    private TextView textViewRu;
-    private TextView textViewDict;
+    public LinearLayout mainControlLayout;
+    private Button btnViewDict;
     private TextView tvWordsCounter;
-    private ImageButton btnPlay;
-    private ImageButton btnStop;
-    private ImageButton btnPause;
-    private ProgressBar progressBar;
-    private CheckBox checkBoxRuSpeak;
-    private ImageView orderPlayIconIV;
-    private static Intent speechIntentService;
-    public static Intent serviceIntent;
-    private SpeechServiceReceiver speechServiceReceiver;
-    private boolean isFirstTime = true;
-    private AppSettings appSettings;
-    private AppData appData;
-    private ArrayList<String> playList = new ArrayList<>();
-    private DataBaseQueries dataBaseQueries;
-    private Locale localeDefault;
+    private ImageView orderPlayView;
+    private ViewPager2 mainViewPager;
 
-    private final String KEY_ENG_TEXT = "eng_text";
-    private final String KEY_RU_TEXT = "ru_text";
-    private final String KEY_CURRENT_DICT = "current_dict";
     private final String KEY_TV_WORDS_COUNTER = "tv_words_counter";
-    private final String KEY_BTN_PLAY_VISIBLE = "btn_play_visible";
-    private final String KEY_BTN_PAUSE_VISIBLE = "btn_pause_visible";
-    private final String KEY_BTN_STOP_VISIBLE = "btn_stop_visible";
-    private final String KEY_PROG_BAR_VISIBLE = "prog_bar_visible";
 
-    private GetTableListFragm getTableListFragm;
-    private FragmentManager fragmentManager;
+    public MainViewModel mainViewModel;
+    private SpeechViewModel speechViewModel;
+    private AdsViewModel adsVM;
+    private final CompositeDisposable composite = new CompositeDisposable();
+    private Word currentWord;
+    private int wordsInterval = Integer.MAX_VALUE;
+
+    @Inject
+    AlarmScheduler scheduler;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState)
     {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.a_navig_main);
+        ANavigMainBinding navBinding = ANavigMainBinding.inflate(getLayoutInflater());
+        View root = navBinding.getRoot();
+        setContentView(root);
 
-        fragmentManager = getSupportFragmentManager();
+        Toolbar toolbar = findViewById(R.id.toolbar_word_editor);
+        setSupportActionBar(toolbar);
+
+        MobileAds.initialize(this);
+
+        NotificationManager nmg = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+        if (nmg != null)
+        {
+            nmg.cancelAll();
+        }
+        scheduler.cancel(AlarmScheduler.REQUEST_CODE, AlarmScheduler.REPEAT_SHOOT_ACTION);
+
+        mainViewModel = new ViewModelProvider(this).get(MainViewModel.class);
+        speechViewModel = new ViewModelProvider(this).get(SpeechViewModel.class);
+        BillingViewModel billingVM = new ViewModelProvider(this).get(BillingViewModel.class);
+        adsVM = new ViewModelProvider(this).get(AdsViewModel.class);
+
+        billingVM.getNoAdsToken().observe(this, t -> {
+            if (t != null && t.isEmpty())
+            {
+                LinearLayout adLayout = findViewById(R.id.adLayout);
+                if (adLayout != null)
+                {
+                    AdView mainBanner = adsVM.getMainBanner();
+                    adLayout.removeAllViews();
+                    adLayout.addView(mainBanner);
+                    mainBanner.loadAd(new AdRequest.Builder().build());
+                }
+            }
+        });
+
+        btnViewDict = findViewById(R.id.btnViewDict);
+        btnViewDictOnClick(btnViewDict);
+        mainViewModel.getCurrentWord().observe(this, word -> {
+            currentWord = word;
+            btnViewDict.setText(currentWord.getDictName());
+        });
+
+        mainViewModel.currentDict.observe(this, dict -> {
+            if (!dict.isEmpty())
+            {
+                btnViewDict.setText(dict);
+            } else
+            {
+                btnViewDict.setText(getString(R.string.text_dictionary));
+            }
+        });
+
+
+        orderPlayView = findViewById(R.id.order_play_icon_iv);
+        orderPlayViewOnClick(orderPlayView);
+        mainViewModel.getOrderPlay().observe(this, order -> {
+            if (order == 0)
+            {
+                mainViewModel.sortWordsList();
+                orderPlayView.setImageResource(R.drawable.ic_repeat_white);
+            }
+            else if (order == 1)
+            {
+                mainViewModel.shuffleWordsList();
+                orderPlayView.setImageResource(R.drawable.ic_shuffle_white);
+            }
+        });
+
+        mainViewPager = findViewById(R.id.mainViewPager);
+        mainViewModel.wordsList.observe(this, entries  -> {
+            if (entries != null && !entries.isEmpty())
+            {
+                MainViewPagerAdapter pagerAdapter = new MainViewPagerAdapter(entries);
+                mainViewPager.setAdapter(pagerAdapter);
+                mainViewPager.setOrientation(ViewPager2.ORIENTATION_HORIZONTAL);
+                for (int i = 0; i < entries.size(); i++)
+                {
+                    int entriesId = entries.get(i).get_id();
+                    Word currentWord = mainViewModel.getCurrentWord().getValue();
+                    if (currentWord != null)
+                    {
+                        int currentId = currentWord.get_id();
+                        if (entriesId >= currentId)
+                        {
+                            mainViewPager.setCurrentItem(i, false);
+                            break;
+                        }
+                    }
+                }
+            }
+        });
+
+        tvWordsCounter = findViewById(R.id.tv_words_counter);
+
+        mainViewPager.registerOnPageChangeCallback(new ViewPager2.OnPageChangeCallback()
+        {
+            private int state = -1;
+            private int position = -1;
+            private Word word;
+
+            @Override
+            public void onPageSelected(int position)
+            {
+                super.onPageSelected(position);
+                final MainViewPagerAdapter adapter = (MainViewPagerAdapter)mainViewPager.getAdapter();
+                if (adapter != null)
+                {
+                    Word item = adapter.getItem(position);
+                    word = item;
+                    mainViewModel.setCurrentWord(item);
+                    int totalWords = adapter.getItemCount();
+                    String concatText = (position + 1) + " / " + totalWords;
+                    tvWordsCounter.setText(concatText);
+                }
+                if (position < mainViewModel.wordListSize() - 1)
+                {
+                    mainViewModel.wordsIsEnded(false);
+                }
+
+            }
+            @Override
+            public void onPageScrollStateChanged(int state)
+            {
+                super.onPageScrollStateChanged(state);
+                this.state = state;
+                if (speechViewModel.isEnSpeech().getValue() != null)
+                {
+                    Boolean isEnSpeech = speechViewModel.isEnSpeech().getValue();
+                    if (state == 1 && isEnSpeech != null && isEnSpeech)
+                    {
+                        if (!currentWord.getEnglish().equals(""))
+                        {
+                            speechViewModel.setSpeechProgressVisibility(View.VISIBLE);
+                        }
+                    }
+                    Integer mControlVisibility = mainViewModel.getMainControlVisibility().getValue();
+                    if (isEnSpeech != null && isEnSpeech && word != null && state == 0 && mControlVisibility != null && mControlVisibility == View.VISIBLE)
+                    {
+                        speechViewModel.doSpeech(word.getEnglish(), Locale.US);
+                    }
+                }
+            }
+
+            @Override
+            public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels)
+            {
+                super.onPageScrolled(position, positionOffset, positionOffsetPixels);
+                int remainder = position % wordsInterval;
+                int testIntervalIndex = position - remainder;
+                mainViewModel.setIntermediateIndex(testIntervalIndex);
+                MainViewPagerAdapter adapter = (MainViewPagerAdapter) mainViewPager.getAdapter();
+                List<Word> list = new ArrayList<>();
+                if (adapter != null)
+                {
+                    boolean condition1 = (state == 2 && remainder == 0 && position != 0 && position > this.position);
+                    boolean condition2 = (state == ViewPager2.SCROLL_STATE_DRAGGING && position == mainViewModel.wordListSize() - 1);
+                    if (condition1)
+                    {
+                        list = adapter.getItems(position - wordsInterval, position - 1);
+                    }
+                    else if (condition2)
+                    {
+                        int lastIndex = adapter.getItemCount() - 1;
+                        int firstIndex = -1;
+                        if (mainViewModel.getIntermediateIndex().getValue() != null)
+                        {
+                            firstIndex = mainViewModel.getIntermediateIndex().getValue();
+                        }
+                        if (firstIndex < 0) firstIndex = 0;
+                        list = adapter.getItems(firstIndex, lastIndex);
+                    }
+                    if ((condition1 || condition2) && list.size() > 1)
+                    {
+                        speechViewModel.stopSpeech();
+                        mainViewModel.setIntermediateIndex(position);
+                        mainViewModel.setMainControlVisibility(View.INVISIBLE);
+                        speechViewModel.setSpeechProgressVisibility(View.INVISIBLE);
+                        Toast.makeText(MainActivity.this, getString(R.string.text_test_knowledge), Toast.LENGTH_LONG).show();
+                        OneOfFiveFragm testFragment = OneOfFiveFragm.newInstance(list);
+                        FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+                        transaction.setCustomAnimations(R.anim.from_right_to_left_anim, R.anim.from_left_to_right_anim);
+                        transaction.replace(R.id.frame_to_page_fragm, testFragment).addToBackStack(null).commit();
+                        mainViewPager.setCurrentItem(position - 1);
+                        return;
+                    }
+                }
+                this.position = position;
+            }
+        });
+
+        mainViewModel.getMainControlVisibility().observe(this, visibility -> {
+            mainControlLayout = findViewById(R.id.main_control_layout);
+            if (mainControlLayout != null)
+            {
+                mainControlLayout.setVisibility(visibility);
+            }
+        });
 
         if (savedInstanceState == null)
         {
@@ -129,27 +306,6 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
             getSupportFragmentManager().beginTransaction().replace(R.id.background_fragment, backgroundFragm).commit();
         }
 
-        Toolbar toolbar = findViewById(R.id.toolbar_word_editor);
-        setSupportActionBar(toolbar);
-
-        appSettings = new AppSettings(this);
-        playList = appSettings.getPlayList();
-        appData = AppData.getInstance();
-        appData.initAllSettings(this);
-
-        initViews();
-
-        // Регистрируем приёмник
-        speechServiceReceiver = new SpeechServiceReceiver();
-        IntentFilter updateIntentFilter = new IntentFilter(SpeechService.ACTION_UPDATE);
-        updateIntentFilter.addCategory(Intent.CATEGORY_DEFAULT);
-        try
-        {
-            registerReceiver(speechServiceReceiver, updateIntentFilter);
-        } catch (Exception e)
-        {
-            e.printStackTrace();
-        }
 
         DrawerLayout drawer = findViewById(R.id.drawer_layout);
         ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
@@ -162,37 +318,146 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
             navigationView.setNavigationItemSelectedListener(this);
         }
 
-        if (savedInstanceState == null && databaseHelper == null)
-        {
-            databaseHelper = new DatabaseHelper(this);
-            databaseHelper.create_db();
-        }
-
         if (savedInstanceState != null)
         {
-            isFirstTime = false;
-            textViewEn.setText(savedInstanceState.getString(KEY_ENG_TEXT));
-            textViewRu.setText(savedInstanceState.getString(KEY_RU_TEXT));
-            textViewDict.setText(savedInstanceState.getString(KEY_CURRENT_DICT));
             tvWordsCounter.setText(savedInstanceState.getString(KEY_TV_WORDS_COUNTER));
-            btnPlay.setVisibility(savedInstanceState.getInt(KEY_BTN_PLAY_VISIBLE));
-            btnStop.setVisibility(savedInstanceState.getInt(KEY_BTN_STOP_VISIBLE));
-            btnPause.setVisibility(savedInstanceState.getInt(KEY_BTN_PAUSE_VISIBLE));
-            progressBar.setVisibility(savedInstanceState.getInt(KEY_PROG_BAR_VISIBLE));
         }
 
-        if (appData.isAdMob())
-        {
-            if (appData.isOnline(this))
+        CheckBox checkBoxEnView = findViewById(R.id.check_box_en_speak);
+        //noinspection CodeBlock2Expr
+        checkBoxEnView.setOnCheckedChangeListener((compoundButton, b) -> {
+            speechViewModel.setEnSpeech(b);
+        });
+        speechViewModel.isEnSpeech().observe(this, checked -> {
+            checkBoxEnView.setChecked(checked);
+            if (!checked)
             {
-                if (savedInstanceState == null)
+                speechViewModel.setSpeechProgressVisibility(View.INVISIBLE);
+            }
+        });
+        speechViewModel.getEnCheckboxEnable().observe(this, checkBoxEnView::setEnabled);
+
+        CheckBox checkBoxRuSpeak = findViewById(R.id.check_box_ru_speak);
+        checkBoxRuSpeak.setOnCheckedChangeListener((compoundButton, b) -> {
+            speechViewModel.setRuSpeech(b);
+            if (b)
+                Toast.makeText(MainActivity.this, R.string.text_ru_speech_on, Toast.LENGTH_SHORT).show();
+            else
+                Toast.makeText(MainActivity.this, R.string.text_ru_speech_off, Toast.LENGTH_SHORT).show();
+        });
+
+        speechViewModel.isRuSpeech().observe(this, checked -> {
+            //noinspection Convert2MethodRef
+            checkBoxRuSpeak.setChecked(checked);
+        });
+        speechViewModel.getRuCheckboxEnable().observe(this, checkBoxRuSpeak::setEnabled);
+
+
+        ProgressBar speechProgress = findViewById(R.id.speechProgress);
+        //noinspection CodeBlock2Expr
+        speechViewModel.getSpeechStartId().observe(MainActivity.this, utteranceId ->
+        {
+            speechViewModel.setSpeechProgressVisibility(View.INVISIBLE);
+        });
+
+        speechViewModel.getSpeechDoneId().observe(MainActivity.this, utteranceId ->
+        {
+            speechViewModel.setSpeechProgressVisibility(View.INVISIBLE);
+            Boolean isRu = speechViewModel.isRuSpeech().getValue();
+            Word word = mainViewModel.getCurrentWord().getValue();
+            if (utteranceId.equals("En") && isRu != null && isRu && word != null)
+            {
+                speechViewModel.doSpeech(word.getTranslate(), Locale.getDefault());
+            }
+        });
+
+        //noinspection CodeBlock2Expr
+        speechViewModel.getSpeechError().observe(this, err -> {
+            speechViewModel.setSpeechProgressVisibility(View.INVISIBLE);
+        });
+
+
+        ImageButton btnSpeak = findViewById(R.id.btn_speak);
+        speechViewModel.getSpeechProgressVisibility().observe(this, v -> {
+            speechProgress.setVisibility(v);
+            if (v == View.VISIBLE)
+                btnSpeak.setVisibility(View.INVISIBLE);
+            else btnSpeak.setVisibility(View.VISIBLE);
+        });
+
+
+        btnSpeak.setOnClickListener(view -> {
+            int position = mainViewPager.getCurrentItem();
+            Boolean isEnSpeech = speechViewModel.isEnSpeech().getValue();
+            Boolean isRuSpeech = speechViewModel.isRuSpeech().getValue();
+            if (mainViewModel.wordsList.getValue() != null)
+            {
+                Word word = mainViewModel.wordsList.getValue().get(position);
+                String enText = word.getEnglish();
+                String ruText = word.getTranslate();
+                if (isEnSpeech != null)
                 {
-                    MobileAds.initialize(this, getString(R.string.admob_app_id));
-                    MainBannerFragment bannerFragment = new MainBannerFragment();
-                    getSupportFragmentManager().beginTransaction().replace(R.id.banner_frame_main, bannerFragment).commit();
+                    speechViewModel.doSpeech(enText, Locale.US);
+                    speechViewModel.setSpeechProgressVisibility(View.VISIBLE);
+                }
+                if (isRuSpeech != null && isRuSpeech)
+                {
+                    Locale localeRu = new Locale(getString(R.string.lang_code_translate));
+                    speechViewModel.doSpeech(ruText, localeRu);
                 }
             }
+        });
+
+        AppCompatImageButton btnReplay = findViewById(R.id.btnReplay);
+        mainViewModel.isEndWordList.observe(this, isEnd -> {
+            if (isEnd)
+            {
+                btnReplay.setVisibility(View.VISIBLE);
+            }
+            else btnReplay.setVisibility(View.INVISIBLE);
+        });
+
+        btnReplay.setOnClickListener( view -> mainViewPager.setCurrentItem(0, true));
+
+    }
+
+    public void testPassed()
+    {
+        if (currentWord != null)
+        {
+            mainViewModel.saveCurrentWordToPref(currentWord);
+            List<Word> wordList = mainViewModel.wordsList.getValue();
+            if (wordList != null)
+            {
+                int index = wordList.indexOf(currentWord);
+                if (index <= wordList.size()-1 && index >=0)
+                {
+                    int i = index + 1;
+                    mainViewPager.setCurrentItem(i, false);
+                }
+                else mainViewPager.setCurrentItem(0, false);
+                mainViewModel.wordsIsEnded(index + 1 == mainViewModel.wordListSize() - 1);
+            }
         }
+        mainViewModel.setMainControlVisibility(View.VISIBLE);
+    }
+
+    public void testFailed(int errors)
+    {
+        if (currentWord != null && errors > 0)
+        {
+            List<Word> wordList = mainViewModel.wordsList.getValue();
+            if (wordList != null && mainViewModel.getIntermediateIndex().getValue() != null)
+            {
+                int newIndex = mainViewModel.getIntermediateIndex().getValue();
+                if (newIndex <= wordList.size()-1 && newIndex >= 0)
+                {
+                    mainViewPager.setCurrentItem(newIndex, true);
+                }
+                else mainViewPager.setCurrentItem(0, true);
+            }
+        }
+        mainViewModel.setMainControlVisibility(View.VISIBLE);
     }
 
     public Action getAction()
@@ -204,145 +469,128 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
     protected void onStart()
     {
         super.onStart();
-        FirebaseAppIndex.getInstance().update();
-        FirebaseUserActions.getInstance().start(getAction());
-        EventBus.getDefault().post(new MainActivityOnStart(serviceIntent));
+        FirebaseAppIndex.getInstance(this).update();
+        FirebaseUserActions.getInstance(this).start(getAction());
     }
 
-    private void initViews()
+    @SuppressWarnings("Convert2Lambda")
+    private void btnViewDictOnClick(Button button)
     {
-        textViewEn = findViewById(R.id.enTextView);
-        textViewRu = findViewById(R.id.ruTextView);
-        textViewDict = findViewById(R.id.textViewDict);
-        tvWordsCounter = findViewById(R.id.tv_words_counter);
-        btnPlay = findViewById(R.id.btn_play);
-        btnStop = findViewById(R.id.btn_stop);
-        if (btnStop != null)
+        button.setOnClickListener(new View.OnClickListener()
         {
-            btnStop.setVisibility(View.GONE);
-        }
-        btnPause = findViewById(R.id.btn_pause);
-        if (btnPause != null)
-        {
-            btnPause.setVisibility(View.GONE);
-        }
-        progressBar = findViewById(R.id.progressBar);
-        checkBoxRuSpeak = findViewById(R.id.check_box_ru_speak);
-        checkBoxRuSpeak.setChecked(appSettings.isEnglishSpeechOnly());
-        switchRuSound_OnCheckedChange();
+            @Override
+            public void onClick(View view)
+            {
+                String buttonText = button.getText().toString();
+                composite.add(mainViewModel.getDictList().subscribeOn(Schedulers.computation())
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe( list -> {
+                            int index = list.indexOf(buttonText);
+                            if (index >= 0)
+                            {
+                                String item = list.remove(index);
+                                list.add(0, item);
+                            }
+                            DictListDialog.Companion.getInstance(list, new DictListDialog.ISelectItemListener()
+                            {
+                                @Override
+                                public void dictListItemOnSelected(@NonNull String dict)
+                                {
+                                    mainViewModel.setWordsList(dict, 1);
+                                    Word word = new Word(1, dict, "", "", 1);
+                                    mainViewModel.saveCurrentWordToPref(word);
+                                    mainViewModel.setCurrentWord(word);
+                                    mainViewModel.setCurrentDict(dict);
+                                }
+                            }).show(getSupportFragmentManager(), DictListDialog.Companion.getTAG());
 
-        orderPlayIconIV = findViewById(R.id.order_play_icon_iv);
+                        }, Throwable::printStackTrace));
+            }
+        });
+    }
+
+    private void orderPlayViewOnClick(ImageView view)
+    {
+        view.setOnClickListener( v -> {
+            Integer order = mainViewModel.getOrderPlay().getValue();
+            if (order != null)
+            {
+                OrderPlayDialog.Companion.getInstance(order, ord -> mainViewModel.setOrderPlay(ord))
+                        .show(getSupportFragmentManager(), OrderPlayDialog.Companion.getTAG());
+            }
+        });
     }
 
     @Override
     protected void onSaveInstanceState(@NonNull Bundle outState)
     {
         super.onSaveInstanceState(outState);
-        outState.putString(KEY_ENG_TEXT, textViewEn.getText().toString());
-        outState.putString(KEY_RU_TEXT, textViewRu.getText().toString());
-        outState.putString(KEY_CURRENT_DICT, textViewDict.getText().toString());
         outState.putString(KEY_TV_WORDS_COUNTER, tvWordsCounter.getText().toString());
-        outState.putInt(KEY_BTN_PLAY_VISIBLE, btnPlay.getVisibility());
-        outState.putInt(KEY_BTN_STOP_VISIBLE, btnStop.getVisibility());
-        outState.putInt(KEY_BTN_PAUSE_VISIBLE, btnPause.getVisibility());
-        outState.putInt(KEY_PROG_BAR_VISIBLE, progressBar.getVisibility());
-    }
-
-    @Override
-    protected void onPause()
-    {
-        super.onPause();
-        appData.saveAllSettings(this);
-        if (getInchesDisplay() < 7)
-        {
-            if (!isActivityOnTop())
-            {
-                speechServiceOnPause();
-            }
-        }
     }
 
     @Override
     protected void onStop()
     {
-        progressBar.setVisibility(View.GONE);
-        FirebaseUserActions.getInstance().end(getAction());
+        FirebaseUserActions.getInstance(this).end(getAction());
         super.onStop();
-    }
-
-    @Override
-    protected void onRestart()
-    {
-        super.onRestart();
     }
 
     @Override
     protected void onResume()
     {
         super.onResume();
-        appSettings = new AppSettings(this);
-        playList = appSettings.getPlayList();
-        appData = AppData.getInstance();
-        appData.initAllSettings(this);
-        appData.setDictNumberChangeListener(this);
-        if (appSettings.getOrderPlay() == 0)
+        if (mainViewModel.getTestInterval().getValue() != null)
         {
-            orderPlayIconIV.setImageResource(R.drawable.ic_repeat_white);
+            wordsInterval = mainViewModel.getTestInterval().getValue();
         }
-        if (appSettings.getOrderPlay() == 1)
+        this.stopService(new Intent(this, LexiconService.class));
+        Boolean isRefresh = AppBus.INSTANCE.isRefresh().getValue();
+        if (isRefresh != null && isRefresh)
         {
-            orderPlayIconIV.setImageResource(R.drawable.ic_shuffle_white);
+            mainViewModel.refreshWordsList();
         }
-        if (playList.size() > 0 && appData.getNdict() < playList.size())
+
+        this.getOnBackPressedDispatcher().addCallback(new OnBackPressedCallback(true)
         {
-            textViewDict.setText(playList.get(appData.getNdict()));
-        }
-        else
-        {
-            textViewDict.setText("");
-        }
-        localeDefault = new Locale(appSettings.getTranslateLang());
-        dataBaseQueries = new DataBaseQueries(this);
-        if (serviceIntent != null)
-        {
-            stopService(serviceIntent);
-        }
+            @Override
+            public void handleOnBackPressed()
+            {
+                getWindow().addFlags(WindowManager.LayoutParams.FLAG_ALLOW_LOCK_WHILE_SCREEN_ON);
+                DrawerLayout drawer = findViewById(R.id.drawer_layout);
+
+                if (drawer != null)
+                {
+                    if (drawer.isDrawerOpen(GravityCompat.START))
+                    {
+                        drawer.closeDrawer(GravityCompat.START);
+                    }
+                }
+                alarmClockEnable(scheduler);
+                finish();
+            }
+        });
     }
 
     @Override
     protected void onDestroy()
     {
+        mainViewModel.saveCurrentWordToPref(currentWord);
         super.onDestroy();
-        if (databaseHelper != null)
-        {
-            databaseHelper.close();
-        }
-        unregisterReceiver(speechServiceReceiver);
-        if (AppData.getInstance().getDisplayVariant() == 1 && serviceIntent != null)
-        {
-            stopService(serviceIntent);
-        }
     }
 
-    @Override
-    public void onBackPressed()
+    private void alarmClockEnable(AlarmScheduler scheduler)
     {
-        getWindow().addFlags(WindowManager.LayoutParams.FLAG_ALLOW_LOCK_WHILE_SCREEN_ON);
-        speechServiceOnPause();
-        appData.saveAllSettings(this);
-        DrawerLayout drawer = findViewById(R.id.drawer_layout);
-
-        if (drawer != null)
+        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(this);
+        String interval = preferences.getString(getString(R.string.key_show_intervals), "0");
+        if (interval != null)
         {
-            if (drawer.isDrawerOpen(GravityCompat.START))
+            int parseInt = Integer.parseInt(interval);
+            if (parseInt != 0)
             {
-                drawer.closeDrawer(GravityCompat.START);
-            } else
-            {
-                super.onBackPressed();
+                //scheduler.scheduleRepeat((parseInt*60*1000), (parseInt*60*1000));
+                scheduler.scheduleOne(parseInt * 60 * 1000);
             }
         }
-
     }
 
     @Override
@@ -356,64 +604,44 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
     public boolean onOptionsItemSelected(MenuItem item)
     {
         int id = item.getItemId();
-
         if (id == R.id.edit_word)
         {
-            if (playList != null && playList.size() > 0)
+            Word word = mainViewModel.getCurrentWord().getValue();
+            if (word != null)
             {
-                if (wordEditorIntent == null)
-                {
-                    wordEditorIntent = new Intent(this, WordEditorActivity.class);
-                }
-                speechServiceOnPause();
-                Bundle bundle = new Bundle();
-                bundle.putString(WordEditorActivity.KEY_EXTRA_DICT_NAME, textViewDict.getText().toString());
-                String text = tvWordsCounter.getText().toString();
-                try
-                {
-                    String[] splitArray = text.split(" ");
-                    if (splitArray.length > 0)
-                    {
-                        bundle.putInt(WordEditorActivity.KEY_ROW_ID, Integer.parseInt(splitArray[0]));
-                    }
-
-                } catch (Exception e)
-                {
-                    e.printStackTrace();
-                }
-                wordEditorIntent.replaceExtras(bundle);
-
-                startActivity(wordEditorIntent);
+                AppBus.INSTANCE.passWord(word); // отправляем слово в WordEditorActivity
             }
-            else
-            {
-                Toast toast = Toast.makeText(this, R.string.no_playlist, Toast.LENGTH_SHORT);
-                toast.setGravity(Gravity.CENTER,0,0);
-                toast.show();
-                if (playListIntent == null)
-                {
-                    playListIntent = new Intent(this, PlayList.class);
-                }
-                startActivity(playListIntent);
-            }
+            Intent intent = new Intent(this, WordEditorActivity.class);
+            startActivity(intent);
         }
         if (id == R.id.edit_speech_data)
         {
-            speechServiceOnPause();
-            Intent speechEditorIntent = new Intent(Intent.ACTION_VIEW);
-            speechEditorIntent.setAction(Settings.ACTION_SETTINGS);
-            startActivity(speechEditorIntent);
+            try
+            {
+                Intent speechEditorIntent = new Intent(Intent.ACTION_VIEW);
+                speechEditorIntent.setAction("com.android.settings.TTS_SETTINGS");
+                speechEditorIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                startActivity(speechEditorIntent);
+            } catch (Exception e)
+            {
+                e.printStackTrace();
+            }
         }
-
         if (id == R.id.cloud_storage)
         {
             StorageFragment2 storageFragment = StorageFragment2.Companion.newInstance();
-            fragmentManager.beginTransaction().replace(R.id.frame_to_page_fragm, storageFragment).addToBackStack(null).commit();
+            getSupportFragmentManager().beginTransaction().replace(R.id.frame_to_page_fragm, storageFragment).addToBackStack(null).commit();
         }
-
         if (id == R.id.menu_item_share)
         {
             new Share().doShare(this);
+        }
+        if (id == R.id.menu_run_migration_db)
+        {
+            DatabaseHelper helper = new DatabaseHelper(this);
+            AppDao roomDb = AppDataBase.Companion.buildDataBase(this).appDao();
+            AppDB appDB = new AppDB(helper, roomDb);
+            appDB.migrateToWordsTable();
         }
 
         return super.onOptionsItemSelected(item);
@@ -423,60 +651,58 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
     public boolean onNavigationItemSelected(@NonNull MenuItem item)
     {
         int id = item.getItemId();
-        speechServiceOnPause();
-
         if (id == R.id.nav_add_word)
         {
-//            if (addWordIntent == null)
-//            {
-//                addWordIntent = new Intent(this,AddWordActivity.class);
-//            }
-//            startActivity(addWordIntent);
             TranslateFragment translateFragm = TranslateFragment.Companion.getInstance("");
-            getSupportFragmentManager().beginTransaction().replace(R.id.frame_to_page_fragm, translateFragm).addToBackStack(null).commit();
-        }
-        else if (id == R.id.nav_add_dict)
-        {
-            dialogAddDict();
+            FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+            transaction.setCustomAnimations(R.anim.from_right_to_left_anim, R.anim.from_left_to_right_anim);
+            transaction.replace(R.id.frame_to_page_fragm, translateFragm).addToBackStack(null).commit();
         }
         else if (id == R.id.nav_delete_dict)
         {
-            getTableListFragm = new GetTableListFragm();
-            fragmentManager.beginTransaction().add(getTableListFragm, GetTableListFragm.TAG).commit();
+            Disposable subscribe = mainViewModel.getDictList()
+                    .subscribeOn(Schedulers.newThread())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .subscribe(list -> {
+                        if (list != null && !list.isEmpty())
+                        {
+                            //noinspection Convert2Lambda
+                            RemoveDictDialog.Companion.getInstance((ArrayList<String>) list, new RemoveDictDialog.IRemoveDictDialogCallback()
+                            {
+                                @Override
+                                public void removeDictDialogButtonClickListener(@NonNull List<String> list)
+                                {
+                                    boolean contains = list.contains(mainViewModel.currentDict.getValue());
+                                    if (contains)
+                                    {
+                                        mainViewModel.resetWordsList();
+                                    }
+                                }
+                            }).show(getSupportFragmentManager(), RemoveDictDialog.TAG);
+                        }
+                    }, Throwable::printStackTrace);
+            composite.add(subscribe);
         }
         else if (id == R.id.nav_edit)
         {
-            if (wordEditorIntent == null)
-            {
-                wordEditorIntent = new Intent(this, WordEditorActivity.class);
-            }
+            Intent intent = new Intent(this, WordEditorActivity.class);
             Bundle bundle = new Bundle();
-            if (playList.size() > 0)
-            {
-                bundle.putString(WordEditorActivity.KEY_EXTRA_DICT_NAME, playList.get(AppData.getInstance().getNdict()));
-                wordEditorIntent.replaceExtras(bundle);
-            }
-            startActivity(wordEditorIntent);
+            bundle.putString(WordEditorActivity.KEY_EXTRA_DICT_NAME, btnViewDict.getText().toString());
+            intent.replaceExtras(bundle);
+            startActivity(intent);
         }
         else if (id == R.id.nav_check_your_self)
         {
-            if (testsIntent == null)
-            {
-                testsIntent = new Intent(this, TestsActivity.class);
-            }
-            startActivity(testsIntent);
-        }
-        else if (id == R.id.nav_play_list)
-        {
-            if (playListIntent == null)
-            {
-                playListIntent = new Intent(this, PlayList.class);
-            }
-            startActivity(playListIntent);
+            getSupportFragmentManager().beginTransaction().replace(R.id.frame_to_page_fragm, TestFragment.Companion.newInstance()).addToBackStack(null).commit();
+//            if (testsIntent == null)
+//            {
+//                testsIntent = new Intent(this, TestsActivity.class);
+//            }
+//            startActivity(testsIntent);
         }
         else if (id == R.id.nav_settings)
         {
-            getFragmentManager().beginTransaction().replace(R.id.frame_to_page_fragm, new SettingsFragment()).addToBackStack(null).commit();
+            getSupportFragmentManager().beginTransaction().replace(R.id.frame_to_page_fragm, new SettingsFragment()).addToBackStack(null).commit();
         }
         else if (id == R.id.nav_evaluate_app)
         {
@@ -491,7 +717,7 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
         else if (id == R.id.nav_about_app)
         {
             AboutAppFragment aboutAppFragment = new AboutAppFragment();
-            fragmentManager.beginTransaction().replace(R.id.frame_to_page_fragm, aboutAppFragment).addToBackStack(null).commit();
+            getSupportFragmentManager().beginTransaction().replace(R.id.frame_to_page_fragm, aboutAppFragment).addToBackStack(null).commit();
         }
         else if (id == R.id.nav_exit)
         {
@@ -500,6 +726,7 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
                 SplashScreenActivity.speech.stop();
                 SplashScreenActivity.speech.shutdown();
             }
+            alarmClockEnable(scheduler);
             this.finish();
         }
 
@@ -511,417 +738,13 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
         return true;
     }
 
-    @Override
-    public void onGetTableListListener(Object object)
-    {
-        getTableListFragm = (GetTableListFragm) fragmentManager.findFragmentByTag("get_table_list");
-        if (getTableListFragm != null)
-        {
-            fragmentManager.beginTransaction().remove(getTableListFragm).commit();
-        }
-        @SuppressWarnings("unchecked") ArrayList<String> arrayList = (ArrayList<String>) object;
-        final ArrayList<String> delete_items = new ArrayList<>();
-        final String[] items = new  String[arrayList.size()];
-        for (int i = 0; i < arrayList.size(); i++)
-        {
-            items[i] = arrayList.get(i);
-        }
-        boolean[]choice = new boolean[items.length];
-        new AlertDialog.Builder(MainActivity.this).setTitle(R.string.title_del_dict)
-                .setMultiChoiceItems(items, choice, new DialogInterface.OnMultiChoiceClickListener()
-                {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which, boolean isChecked)
-                    {
-                        delete_items.add(items[which]);
-                    }
-                }).setPositiveButton(R.string.button_text_delete, new DialogInterface.OnClickListener()
-        {
-            @Override
-            public void onClick(DialogInterface dialog, int which)
-            {
-                if(delete_items.size() <= 0)    return;
-                new AlertDialog.Builder(MainActivity.this).setTitle(R.string.dialog_are_you_sure)
-                        .setPositiveButton(R.string.button_text_yes, new DialogInterface.OnClickListener()
-                        {
-                            @Override
-                            public void onClick(DialogInterface dialog, int which)
-                            {
-                                boolean result = false;
-                                for (String item : delete_items)
-                                {
-                                    try
-                                    {
-                                        dataBaseQueries = new DataBaseQueries(MainActivity.this);
-                                        result = dataBaseQueries.deleteTableFromDbSync(item);
-                                        appSettings.removeItemFromPlayList(item);
-                                    } catch (Exception e)
-                                    {
-                                        e.printStackTrace();
-                                    }
-                                }
-                                if (result)
-                                {
-                                    Toast.makeText(MainActivity.this, R.string.msg_selected_dict_removed, Toast.LENGTH_LONG).show();
-                                }
-                            }
-                        }).setNegativeButton(R.string.button_text_no, null).create().show();
-            }
-        }).setNegativeButton(R.string.button_text_cancel,null).create().show();
-    }
-
-    private void dialogAddDict()
-    {
-        final View view = getLayoutInflater().inflate(R.layout.a_dialog_add_dict, new LinearLayout(this), false);
-        final EditText editText = view.findViewById(R.id.dialog_add_dict);
-        editText.setInputType(InputType.TYPE_TEXT_FLAG_CAP_WORDS);
-        new AlertDialog.Builder(this).setTitle(R.string.title_new_dictionary).setIcon(R.drawable.icon_book)
-                .setPositiveButton(R.string.btn_text_add, new DialogInterface.OnClickListener()
-                {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which)
-                    {
-
-                        String dictName = editText.getText().toString();
-                        if (!dictName.equals(""))
-                        {
-                            try
-                            {
-                                dataBaseQueries = new DataBaseQueries(MainActivity.this);
-                                dataBaseQueries.addTableToDbSync(dictName);
-                            } catch (Exception e)
-                            {
-                                e.printStackTrace();
-                            }
-                            Toast toast = Toast.makeText(getApplicationContext(), getString(R.string.text_added_new_dict)+dictName, Toast.LENGTH_LONG);
-                            toast.setGravity(Gravity.CENTER, 0, 0);
-                            toast.show();
-                        }
-                    }
-                })
-                .setNegativeButton(R.string.btn_text_cancel, null)
-                .setView(view).create().show();
-
-        editText.addTextChangedListener(new TextWatcher()
-        {
-            @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after){}
-
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count)
-            {
-                String str = s.toString();
-                for (int i = 0; i < str.length(); i++)
-                {
-                    int char_first = str.codePointAt(0);
-                    if ((char_first >= 33 && char_first <= 64) || (char_first >= 91 && char_first <= 96) ||
-                            (char_first >= 123 && char_first <= 126))
-                    {
-                        String str_wrong = str.substring(i);
-                        editText.setText(str.replace(str_wrong,""));
-                    }
-                    if ((str.codePointAt(i) >= 33 && str.codePointAt(i) <= 47) || (str.codePointAt(i) >= 58 && str.codePointAt(i) <= 64) ||
-                            (str.codePointAt(i) >= 91 && str.codePointAt(i) <= 96) || (str.codePointAt(i) >= 123 && str.codePointAt(i) <= 126))
-                    {
-                        String str_wrong = str.substring(i);
-                        editText.setText(str.replace(str_wrong,""));
-                    }
-                }
-            }
-
-            @Override
-            public void afterTextChanged(Editable s){}
-        });
-    }
-
-    public void btnPlayClick(View view)
-    {
-        playList = appSettings.getPlayList();
-        if (playList.size() > 0)
-        {
-            if (isFirstTime)
-            {
-                Toast toast = Toast.makeText(this, R.string.message_about_start,Toast.LENGTH_LONG);
-                toast.setGravity(Gravity.TOP,0,0);
-                toast.show();
-            }
-
-            speechIntentService = new Intent(this, SpeechService.class);
-            stopService(speechIntentService);
-            speechIntentService.putExtra(getString(R.string.key_play_order), appSettings.getOrderPlay());
-            speechIntentService.putExtra(getString(R.string.is_one_time), false);
-            startService(speechIntentService);
-
-            btnPlay.setVisibility(View.INVISIBLE);
-            btnStop.setVisibility(View.VISIBLE);
-            btnPause.setVisibility(View.VISIBLE);
-            textViewRu.setText(null);
-        } else
-        {
-            Toast toast = Toast.makeText(this, R.string.no_playlist, Toast.LENGTH_SHORT);
-            toast.setGravity(Gravity.CENTER,0,0);
-            toast.show();
-            if (playListIntent == null)
-            {
-                playListIntent = new Intent(this, PlayList.class);
-            }
-            startActivity(playListIntent);
-        }
-        progressBar.setVisibility(View.VISIBLE);
-
-    }
-    public void btnPauseClick(View view)
-    {
-        Toast toast = Toast.makeText(this, R.string.message_about_pause,Toast.LENGTH_SHORT);
-        toast.show();
-        speechServiceOnPause();
-    }
-
-    private void speechServiceOnPause()
-    {
-        appData.setPause(true);
-        if (speechIntentService != null)
-        {
-            stopService(speechIntentService);
-            SpeechService.stopIntentService();
-        }
-        btnPause.setVisibility(View.GONE);
-        btnPlay.setVisibility(View.VISIBLE);
-        btnStop.setVisibility(View.VISIBLE);
-        progressBar.setVisibility(View.GONE);
-    }
-
-    public void btnStopClick(View view)
-    {
-        speechServiceOnStop();
-    }
-
-    private void speechServiceOnStop()
-    {
-        if (speechIntentService != null)
-        {
-            stopService(speechIntentService);
-        }
-        textViewEn.setText(null);
-        textViewRu.setText(null);
-        btnPlay.setVisibility(View.VISIBLE);
-        btnStop.setVisibility(View.GONE);
-        btnPause.setVisibility(View.GONE);
-        progressBar.setVisibility(View.GONE);
-
-        appData.setNdict(0);
-        appData.setNword(1);
-    }
-
-    public void btnNextBackClick(View view)
-    {
-        playList = appSettings.getPlayList();
-        if (playList == null || playList.size() == 0)
-        {
-            Toast toast = Toast.makeText(this, R.string.no_playlist, Toast.LENGTH_SHORT);
-            toast.setGravity(Gravity.CENTER,0,0);
-            toast.show();
-            if (playListIntent == null)
-            {
-                playListIntent = new Intent(this, PlayList.class);
-            }
-            startActivity(playListIntent);
-        }
-        speechServiceOnPause();
-        if (SplashScreenActivity.speech != null)
-        {
-            SplashScreenActivity.speech.stop();
-        }
-        int id = view.getId();
-        if (id == R.id.btn_next)
-        {
-            getNext();
-        }
-
-        if (id == R.id.btn_previous)
-        {
-            getPrevious();
-        }
-    }
-
-    private void getNext()
-    {
-        if (playList != null && playList.size() == 0)
-        {
-            speechServiceOnStop();
-            return;
-        }
-
-        appData.getNextNword(this, new AppData.IGetWordListerner()
-        {
-            @Override
-            public void getWordComplete(ArrayList<DataBaseEntry> entries, Integer[] dictSize)
-            {
-                if (entries.size() > 0)
-                {
-                    DataBaseEntry dataBaseEntry = entries.get(0);
-                    textViewEn.setText(dataBaseEntry.getEnglish());
-                    textViewRu.setText(dataBaseEntry.getTranslate());
-                    textViewDict.setText(playList.get(appData.getNdict()));
-                    String concatText = (dataBaseEntry.getRowId() + "").concat(" / ").concat(Integer.toString(dictSize[1])).concat("  " + getString(R.string.text_studied) + " " + dictSize[2]);
-                    tvWordsCounter.setText(concatText);
-
-                    final HashMap<String, String> hashMap = new HashMap<>();
-                    hashMap.put(TextToSpeech.Engine.KEY_PARAM_UTTERANCE_ID, "main_activity");
-                    try
-                    {
-                        SplashScreenActivity.speech.setLanguage(Locale.US);
-                    } catch (Exception e)
-                    {
-                        return;
-                    }
-                    SplashScreenActivity.speech.setOnUtteranceProgressListener(null);
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
-                    {
-                        SplashScreenActivity.speech.speak(textViewEn.getText().toString(), TextToSpeech.QUEUE_ADD, null, hashMap.get(TextToSpeech.Engine.KEY_PARAM_UTTERANCE_ID));
-                    } else
-                    {
-                        SplashScreenActivity.speech.speak(textViewEn.getText().toString(), TextToSpeech.QUEUE_ADD, hashMap);
-                    }
-                    SplashScreenActivity.speech.setOnUtteranceProgressListener(new UtteranceProgressListener()
-                    {
-                        @Override
-                        public void onStart(String utteranceId)
-                        {
-
-                        }
-
-                        @Override
-                        public void onDone(String utteranceId)
-                        {
-                            if (utteranceId.equals("main_activity") && appSettings.isEnglishSpeechOnly())
-                            {
-                                SplashScreenActivity.speech.setLanguage(localeDefault);
-                                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
-                                {
-                                    SplashScreenActivity.speech.speak(textViewRu.getText().toString(), TextToSpeech.QUEUE_ADD, null, hashMap.get(TextToSpeech.Engine.KEY_PARAM_UTTERANCE_ID));
-                                } else
-                                {
-                                    SplashScreenActivity.speech.speak(textViewRu.getText().toString(), TextToSpeech.QUEUE_ADD, hashMap);
-                                }
-                            }
-                            SplashScreenActivity.speech.setOnUtteranceProgressListener(null);
-                        }
-
-                        @Override
-                        public void onError(String utteranceId)
-                        {
-
-                        }
-                    });
-                }
-            }
-        });
-    }
-
-    private void getPrevious()
-    {
-        if (playList != null && playList.size() == 0)
-        {
-            speechServiceOnStop();
-            return;
-        }
-
-        appData.getPreviousNword(this, new AppData.IGetWordListerner()
-        {
-            @Override
-            public void getWordComplete(ArrayList<DataBaseEntry> entries, Integer[] dictSize)
-            {
-                if (entries.size() > 0)
-                {
-                    DataBaseEntry dataBaseEntry = entries.get(0);
-
-                    textViewEn.setText(dataBaseEntry.getEnglish());
-                    textViewRu.setText(dataBaseEntry.getTranslate());
-                    textViewDict.setText(playList.get(appData.getNdict()));
-                    String concatText = (dataBaseEntry.getRowId() + "").concat(" / ").concat(Integer.toString(dictSize[1])).concat("  " + getString(R.string.text_studied) + " " + dictSize[2]);
-                    tvWordsCounter.setText(concatText);
-
-                    final HashMap<String, String> hashMap = new HashMap<>();
-                    hashMap.put(TextToSpeech.Engine.KEY_PARAM_UTTERANCE_ID, "main_activity");
-                    try
-                    {
-                        SplashScreenActivity.speech.setLanguage(Locale.US);
-                    } catch (Exception e)
-                    {
-                        return;
-                    }
-                    SplashScreenActivity.speech.setOnUtteranceProgressListener(null);
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
-                    {
-                        SplashScreenActivity.speech.speak(textViewEn.getText().toString(), TextToSpeech.QUEUE_ADD, null, hashMap.get(TextToSpeech.Engine.KEY_PARAM_UTTERANCE_ID));
-                    } else
-                    {
-                        SplashScreenActivity.speech.speak(textViewEn.getText().toString(), TextToSpeech.QUEUE_ADD, hashMap);
-                    }
-                    SplashScreenActivity.speech.setOnUtteranceProgressListener(new UtteranceProgressListener()
-                    {
-                        @Override
-                        public void onStart(String utteranceId)
-                        {
-
-                        }
-
-                        @Override
-                        public void onDone(String utteranceId)
-                        {
-                            if (utteranceId.equals("main_activity") && appSettings.isEnglishSpeechOnly())
-                            {
-                                SplashScreenActivity.speech.setLanguage(localeDefault);
-                                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
-                                {
-                                    SplashScreenActivity.speech.speak(textViewRu.getText().toString(), TextToSpeech.QUEUE_ADD, null, hashMap.get(TextToSpeech.Engine.KEY_PARAM_UTTERANCE_ID));
-                                } else
-                                {
-                                    SplashScreenActivity.speech.speak(textViewRu.getText().toString(), TextToSpeech.QUEUE_ADD, hashMap);
-                                }
-                            }
-                            SplashScreenActivity.speech.setOnUtteranceProgressListener(null);
-                        }
-
-                        @Override
-                        public void onError(String utteranceId)
-                        {
-
-                        }
-                    });
-                }
-            }
-        });
-    }
-
-    public void switchRuSound_OnCheckedChange()
-    {
-        checkBoxRuSpeak.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener()
-        {
-            @Override
-            public void onCheckedChanged(CompoundButton compoundButton, boolean isChecked)
-            {
-                if (isChecked)
-                {
-                    appSettings.setEnglishSpeechOnly(true);
-                    SpeechService.setEnglishOnly(appSettings.isEnglishSpeechOnly());
-                    Toast.makeText(MainActivity.this, R.string.text_ru_speech_on,Toast.LENGTH_SHORT).show();
-                }
-                else
-                {
-                    appSettings.setEnglishSpeechOnly(false);
-                    SpeechService.setEnglishOnly(appSettings.isEnglishSpeechOnly());
-                    Toast.makeText(MainActivity.this, R.string.text_ru_speech_off,Toast.LENGTH_SHORT).show();
-                }
-            }
-        });
-    }
+    public static Boolean isActivityRunning = false;
 
     @Override
     public void onAttachedToWindow()
     {
         super.onAttachedToWindow();
+        isActivityRunning = true;
     }
 
     @Override
@@ -930,128 +753,21 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
         super.onDetachedFromWindow();
         SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(this);
         boolean isUseService = preferences.getBoolean("service", true);
-        if (appSettings.getPlayList().size() > 0 && isUseService)
+        if (isUseService)
         {
-            if (serviceIntent == null)
-            {
-                serviceIntent = new Intent(this, LexiconService.class);
-            }
+            Intent serviceIntent = new Intent(this, LexiconService.class);
             startService(serviceIntent);
         }
+        isActivityRunning = false;
     }
 
-    // TODO: ActivityManager.RunningAppProcessInfo Проверка, что активити находится на верху стека
-    public boolean isActivityOnTop()
+    public void testIntervalOnChange(int value)
     {
-        final ActivityManager activityManager = (ActivityManager) getSystemService(Service.ACTIVITY_SERVICE);
-        List<ActivityManager.RunningAppProcessInfo> runningAppProcesses = null;
-        if (activityManager != null)
-        {
-            runningAppProcesses = activityManager.getRunningAppProcesses();
-        }
-        if (runningAppProcesses != null && runningAppProcesses.size() > 0)
-        {
-            String processName = runningAppProcesses.get(0).processName;
-            String packageName = getApplicationInfo().packageName;
-            return processName.equals(packageName);
-        }
-        return false;
+        wordsInterval = value;
     }
 
-    public void btnSpeak_OnClick(View view)
-    {
-        speechServiceOnPause();
-        final HashMap<String, String> hashMap = new HashMap<>();
-        hashMap.put(TextToSpeech.Engine.KEY_PARAM_UTTERANCE_ID, "btn_speak_onclick");
-        try
-        {
-            SplashScreenActivity.speech.setLanguage(Locale.US);
-        } catch (Exception e)
-        {
-            return;
-        }
-        SplashScreenActivity.speech.setOnUtteranceProgressListener(null);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
-        {
-            SplashScreenActivity.speech.speak(textViewEn.getText().toString(), TextToSpeech.QUEUE_ADD, null, hashMap.get(TextToSpeech.Engine.KEY_PARAM_UTTERANCE_ID));
-        } else
-        {
-            SplashScreenActivity.speech.speak(textViewEn.getText().toString(), TextToSpeech.QUEUE_ADD, hashMap);
-        }
-        SplashScreenActivity.speech.setOnUtteranceProgressListener(new UtteranceProgressListener()
-        {
-            @Override
-            public void onStart(String utteranceId)
-            {
 
-            }
 
-            @Override
-            public void onDone(String utteranceId)
-            {
-                if (utteranceId.equals("btn_speak_onclick") && appSettings.isEnglishSpeechOnly())
-                {
-                    SplashScreenActivity.speech.setLanguage(localeDefault);
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
-                    {
-                        SplashScreenActivity.speech.speak(textViewRu.getText().toString(), TextToSpeech.QUEUE_ADD, null, hashMap.get(TextToSpeech.Engine.KEY_PARAM_UTTERANCE_ID));
-                    } else
-                    {
-                        SplashScreenActivity.speech.speak(textViewRu.getText().toString(), TextToSpeech.QUEUE_ADD, hashMap);
-                    }
-                }
-                SplashScreenActivity.speech.setOnUtteranceProgressListener(null);
-            }
-
-            @Override
-            public void onError(String s)
-            {
-
-            }
-        });
-    }
-
-    private double getInchesDisplay()
-    {
-        double screenInches;
-        DisplayMetrics displayMetrics = new DisplayMetrics();
-        getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
-        int widthPixels = displayMetrics.widthPixels;
-        int heightPixels = displayMetrics.heightPixels;
-        int densityDpi = displayMetrics.densityDpi;
-        double width = (double) widthPixels/(double) densityDpi;
-        double height = (double) heightPixels/(double) densityDpi;
-        double x = Math.pow(width,2);
-        double y = Math.pow(height,2);
-        screenInches = Math.sqrt(x+y);
-        return screenInches;
-    }
-
-    @Override
-    public void dictNumberOnChanged(int ndict)
-    {
-
-    }
-
-    public class SpeechServiceReceiver extends BroadcastReceiver
-    {
-        @Override
-        public void onReceive(Context context, Intent intent)
-        {
-            String updateEN = intent.getStringExtra(SpeechService.EXTRA_KEY_EN);
-            String updateRU = intent.getStringExtra(SpeechService.EXTRA_KEY_RU);
-            String updateDict = intent.getStringExtra(SpeechService.EXTRA_KEY_DICT);
-            String nword = intent.getStringExtra(SpeechService.EXTRA_KEY_WORDS_COUNTER);
-            textViewEn.setText(updateEN);
-            textViewRu.setText(updateRU);
-            textViewDict.setText(updateDict);
-            tvWordsCounter.setText(nword);
-            if (!textViewEn.getText().equals(""))
-            {
-                progressBar.setVisibility(View.GONE);
-            }
-        }
-    }
 
 }
 
