@@ -6,12 +6,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.google.android.gms.ads.AdListener;
-import com.google.android.gms.ads.AdRequest;
-import com.google.android.gms.ads.AdView;
-import com.google.android.gms.ads.MobileAds;
 import com.myapp.lexicon.R;
-import com.myapp.lexicon.settings.AppData;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -44,27 +39,27 @@ public class ModalBannerFragment extends Fragment
 
         if (savedInstanceState == null)
         {
-            final AdView bannerView = view.findViewById(R.id.adView_main);
-            bannerView.setVisibility(View.GONE);
-            MobileAds.initialize(getActivity(), getString(R.string.modal_bottom_banner));
-            AdRequest adRequest;
-            if (AppData.getInstance().testDeviceEnabled())
-            {
-                adRequest = new AdRequest.Builder().addTestDevice(AppData.getInstance().getTestDeviceID()).build();
-            } else
-            {
-                adRequest = new AdRequest.Builder().build();
-            }
-            bannerView.loadAd(adRequest);
-            bannerView.setAdListener(new AdListener()
-            {
-                @Override
-                public void onAdLoaded()
-                {
-                    super.onAdLoaded();
-                    bannerView.setVisibility(View.VISIBLE);
-                }
-            });
+//            final AdView bannerView = view.findViewById(R.id.adView_main);
+//            bannerView.setVisibility(View.GONE);
+//            MobileAds.initialize(getActivity(), getString(R.string.modal_bottom_banner));
+//            AdRequest adRequest;
+//            if (AppData.getInstance().testDeviceEnabled())
+//            {
+//                adRequest = new AdRequest.Builder().addTestDevice(AppData.getInstance().getTestDeviceID()).build();
+//            } else
+//            {
+//                adRequest = new AdRequest.Builder().build();
+//            }
+//            bannerView.loadAd(adRequest);
+//            bannerView.setAdListener(new AdListener()
+//            {
+//                @Override
+//                public void onAdLoaded()
+//                {
+//                    super.onAdLoaded();
+//                    bannerView.setVisibility(View.VISIBLE);
+//                }
+//            });
         }
 
         return view;
